+++ controllers/admin/user.controller.ts
@@ -3,12 +3,15 @@ import AccountModel from "../../models/account.model";
 import moment from "moment";
 import md5 from "md5";
 import { systemConfig } from "../../config/adminPrefix";
+import RoleModel from "../../models/roles.model";
 
 
 // [GET] admin/user
 export const index = async ( req: Request, res: Response ) => {
 
-    const userInfo = await AccountModel.find();
+    const userInfo = await AccountModel.find({
+        _id: { $ne: res.locals.account._id }
+    });
 
     const user = userInfo.map(item => ({
         ...item.toObject(), 
@@ -30,12 +33,24 @@ export const detail = async (req: Request, res: Response) => {
     const idUser = req.params.id;
 
     const userInfo = await AccountModel.findOne({ _id: idUser });
-
     userInfo.dateBirth = moment(userInfo.dateBirth).format("YYYY-MM-DD");
+    
+    const role = await RoleModel.findOne( {_id: userInfo.role_id} ).select("permission").lean();
 
+    let id_role = null;
+
+    if(res.locals.roles.permission.includes("all")) {
+
+        id_role = await RoleModel.find({ title: { $ne: "Admin" } }).select("_id");
+
+    };
+    
+    
     res.render("admin/pages/user/edit", {
         title: "Admin | UserInfo",
-        userInfo
+        userInfo,
+        role: role ? role : "",
+        id_role: id_role
     });
 
 };
@@ -43,29 +58,41 @@ export const detail = async (req: Request, res: Response) => {
 // [PATCH] admin/user/detail/:id
 export const detailPatch = async (req: Request, res: Response) => {
 
-    await AccountModel.updateOne(
-        {
-          _id: req.params.id,
-          deleted: false,
-        },
-        req.body
-    );
+    if (res.locals.roles.permission.includes("crud-user")) {
 
-    res.redirect(`/${systemConfig.prefixAdmin}/user`);
+        await AccountModel.updateOne(
+            {
+              _id: req.params.id,
+              deleted: false,
+            },
+            req.body
+        );
+    
+        res.redirect(`/${systemConfig.prefixAdmin}/user`);
+
+    } else {
+        res.redirect(`/${systemConfig.prefixAdmin}/`);
+    };
 
 };
 
 // [DELETE] admin/user/delete/:id
 export const deleteUser = async (req: Request, res: Response) => {
 
-    const id = req.params.id;
+    if (res.locals.roles.permission.includes("crud-user")) {
 
-    await AccountModel.deleteOne({
-        _id: id
-    });
+        const id = req.params.id;
 
-    res.json({
-        code: 200
-    });
+        await AccountModel.deleteOne({
+            _id: id
+        });
+    
+        res.json({
+            code: 200
+        });
+
+    } else {
+        res.redirect(`/${systemConfig.prefixAdmin}/`);
+    };
 
 };
